// src/workdone/dto/create-workdone.dto/create-workdone.dto.ts
/* eslint-disable prettier/prettier */
export class CreateWorkdoneDto {
        "id" : number;
        "scheduled date" : number;
        "work done": string;
        "observations": string;
        "hard required": string;
        "resolution": number;
    scheduled_date: number;
    work_done: string;
    hard_required: string;

}