// src/workdone/dto/create-workdone.dto/update-workdone.dto.ts
/* eslint-disable prettier/prettier */
export class UpdateWorkdoneDto {
    "id" : number;
    "scheduled date" : number;
    "work done": string;
    "observations": string;
    "hard required": string;
    "resolution": number;
    scheduled_date: number;
    work_done: string;
    hard_required: string;
}